chore: checking in refactor WIP
