+++ Cargo.toml
@@ -45,7 +45,6 @@ envfile = "0.2.1"
 env_logger = "0.10.2"
 fast-math = "0.1"
 getrandom = { version = "0.1.16", features = ["custom"] }
-
 hex = "0.4.3"
 log = "0.4.22"
 matches = "0.1.10"
@@ -95,6 +94,8 @@ jito-tip-router-client = { path = "./clients/rust/jito_tip_router", version = "0
 jito-tip-router-core = { path = "./core", version = "=0.0.1" }
 jito-tip-router-program = { path = "./program", version = "=0.0.1" }
 jito-tip-router-shank-cli = { path = "./shank_cli", version = "=0.0.1" }
+jito-tip-distribution = { git = "https://github.com/exo-tech-xyz/jito-programs.git", rev = "c21f7a2d84cba6113d2884328ee9753ab8c368f6", features = ["cpi"] }
+jito-tip-payment = { git = "https://github.com/exo-tech-xyz/jito-programs.git", rev = "c21f7a2d84cba6113d2884328ee9753ab8c368f6", features = ["cpi"] }
 jito-bytemuck = { git = "https://github.com/jito-foundation/restaking.git", rev = "eaf88e7e5ca2845fe3108c4fc4a06a25f9a8514d" }
 jito-account-traits-derive = { git = "https://github.com/jito-foundation/restaking.git", rev = "eaf88e7e5ca2845fe3108c4fc4a06a25f9a8514d" }
 jito-jsm-core = { git = "https://github.com/jito-foundation/restaking.git", rev = "eaf88e7e5ca2845fe3108c4fc4a06a25f9a8514d" }
