+++ meta_merkle_tree/Cargo.toml
@@ -10,6 +10,7 @@ edition = { workspace = true }
 readme = { workspace = true }
 
 [dependencies]
+anchor-lang = {workspace = true}
 borsh = { workspace = true }
 bytemuck = { workspace = true }
 fast-math = { workspace = true }
@@ -19,6 +20,8 @@ jito-jsm-core = { workspace = true }
 jito-restaking-core = { workspace = true }
 jito-restaking-sdk = { workspace = true }
 jito-tip-distribution-sdk = { workspace = true }
+jito-tip-distribution = { workspace = true }
+jito-tip-payment = { workspace = true }
 jito-vault-core = { workspace = true }
 jito-vault-sdk = { workspace = true }
 rand = { workspace = true }
