+++ tip-router-operator-cli/Cargo.toml
@@ -16,8 +16,8 @@ futures = "0.3"
 gethostname = "0.4"
 im = "15.1"
 itertools = "0.11"
-jito-tip-distribution = { git = "https://github.com/exo-tech-xyz/jito-programs.git", rev = "6e5cbce5e2866da0585ef4fa35ee1168fce4556f", features = ["cpi"] }
-jito-tip-payment = { git = "https://github.com/exo-tech-xyz/jito-programs.git", rev = "6e5cbce5e2866da0585ef4fa35ee1168fce4556f", features = ["cpi"] }
+jito-tip-distribution.workspace = true
+jito-tip-payment.workspace = true
 meta-merkle-tree.workspace = true
 solana-account-decoder.workspace = true
 solana-accounts-db.workspace = true
