+++ tip-router-operator-cli/src/lib.rs
@@ -78,6 +78,7 @@ use {
     anchor_lang::prelude::*,
     std::result::Result,
     ellipsis_client::EllipsisClient,
+    meta_merkle_tree::generated_merkle_tree::TreeNode
 };
 
 #[derive(Clone, Deserialize, Serialize, Debug)]
@@ -147,65 +148,6 @@ fn emit_inconsistent_tree_node_amount_dp(
     }
 }
 
-impl GeneratedMerkleTreeCollection {
-    pub async fn new_from_stake_meta_collection(
-        stake_meta_coll: StakeMetaCollection,
-        protocol_fee_bps: u16,
-    ) -> Result<GeneratedMerkleTreeCollection, MerkleRootGeneratorError> {
-        let (config_pda, _) = Pubkey::find_program_address(
-            &[CONFIG_ACCOUNT_SEED],
-            &stake_meta_coll.tip_distribution_program_id,
-        );
-
-        let generated_merkle_trees = stake_meta_coll.stake_metas
-            .into_iter()
-            .filter(|stake_meta| stake_meta.maybe_tip_distribution_meta.is_some())
-            .filter_map(|stake_meta| {
-                // Use the helper function to create tree nodes
-                let mut tree_nodes = match TreeNode::vec_from_stake_meta(
-                    &stake_meta,
-                    protocol_fee_bps,
-                    &stake_meta_coll.tip_distribution_program_id, // Pass the program ID
-                ) {
-                    Err(e) => return Some(Err(e)),
-                    Ok(maybe_tree_nodes) => maybe_tree_nodes,
-                }?;
-
-                // Create merkle tree and add proofs
-                let hashed_nodes: Vec<[u8; 32]> = tree_nodes
-                    .iter()
-                    .map(|n| n.hash().to_bytes())
-                    .collect();
-
-                let tip_distribution_meta = stake_meta.maybe_tip_distribution_meta.unwrap();
-
-                let merkle_tree = MerkleTree::new(&hashed_nodes[..], true);
-                let max_num_nodes = tree_nodes.len() as u64;
-
-                for (i, tree_node) in tree_nodes.iter_mut().enumerate() {
-                    tree_node.proof = Some(get_proof(&merkle_tree, i));
-                }
-
-                Some(Ok(GeneratedMerkleTree {
-                    max_num_nodes,
-                    tip_distribution_account: tip_distribution_meta.tip_distribution_pubkey,
-                    merkle_root_upload_authority: tip_distribution_meta.merkle_root_upload_authority,
-                    merkle_root: *merkle_tree.get_root().unwrap(),
-                    tree_nodes,
-                    max_total_claim: tip_distribution_meta.total_tips,
-                }))
-            })
-            .collect::<Result<Vec<_>, MerkleRootGeneratorError>>()?;
-
-        Ok(GeneratedMerkleTreeCollection {
-            generated_merkle_trees,
-            bank_hash: stake_meta_coll.bank_hash,
-            epoch: stake_meta_coll.epoch,
-            slot: stake_meta_coll.slot,
-        })
-    }
-}
-
 pub fn get_proof(merkle_tree: &MerkleTree, i: usize) -> Vec<[u8; 32]> {
     let mut proof = Vec::new();
     let path = merkle_tree.find_path(i).expect("path to index");
@@ -244,159 +186,6 @@ fn derive_tip_payment_pubkeys(program_id: &Pubkey) -> TipPaymentPubkeys {
             tip_pda_6,
             tip_pda_7
         ],
-    }
-}
-
-#[derive(Clone, Eq, Debug, Hash, PartialEq, Deserialize, Serialize)]
-pub struct TreeNode {
-    /// The stake account entitled to redeem.
-    #[serde(with = "pubkey_string_conversion")]
-    pub claimant: Pubkey,
-
-    /// Pubkey of the ClaimStatus PDA account, this account should be closed to reclaim rent.
-    #[serde(with = "pubkey_string_conversion")]
-    pub claim_status_pubkey: Pubkey,
-
-    /// Bump of the ClaimStatus PDA account
-    pub claim_status_bump: u8,
-
-    #[serde(with = "pubkey_string_conversion")]
-    pub staker_pubkey: Pubkey,
-
-    #[serde(with = "pubkey_string_conversion")]
-    pub withdrawer_pubkey: Pubkey,
-
-    /// The amount this account is entitled to.
-    pub amount: u64,
-
-    /// The proof associated with this TreeNode
-    pub proof: Option<Vec<[u8; 32]>>,
-}
-
-impl TreeNode {
-    fn vec_from_stake_meta(
-        stake_meta: &StakeMeta,
-        protocol_fee_bps: u16,
-        tip_distribution_program_id: &Pubkey,
-    ) -> Result<Option<Vec<TreeNode>>, MerkleRootGeneratorError> {
-        
-        if let Some(tip_distribution_meta) = stake_meta.maybe_tip_distribution_meta.as_ref() {
-            let protocol_fee_amount = u128::div_ceil(
-                tip_distribution_meta.total_tips as u128 * protocol_fee_bps as u128,
-                10_000,
-            );
-            let protocol_fee_amount = u64::try_from(protocol_fee_amount)
-                .map_err(|_| MerkleRootGeneratorError::CheckedMathError)?;
-
-            let validator_amount = u128::div_ceil(
-                tip_distribution_meta.total_tips as u128 * tip_distribution_meta.validator_fee_bps as u128,
-                10_000,
-            );
-            let validator_amount = u64::try_from(validator_amount)
-                .map_err(|_| MerkleRootGeneratorError::CheckedMathError)?;
-
-            let remaining_total_rewards = tip_distribution_meta.total_tips
-                .checked_sub(protocol_fee_amount)
-                .and_then(|v| v.checked_sub(validator_amount))
-                .ok_or(MerkleRootGeneratorError::CheckedMathError)?;
-
-            println!("Protocol Fee Amount: {}", protocol_fee_amount);
-            println!("Validator Amount: {}", validator_amount);
-            println!("Remaining Total Rewards: {}", remaining_total_rewards);
-
-            let (protocol_fee_recipient, _) = Pubkey::find_program_address(
-                &[b"protocol_fee", &(0u64).to_le_bytes()],
-                tip_distribution_program_id,
-            );
-
-            let (protocol_claim_status_pubkey, protocol_claim_status_bump) = Pubkey::find_program_address(
-                &[
-                    ClaimStatus::SEED,
-                    &protocol_fee_recipient.to_bytes(),
-                    &tip_distribution_meta.tip_distribution_pubkey.to_bytes(),
-                ],
-                &JitoTipDistribution::id(),
-            );
-
-            let mut tree_nodes = vec![TreeNode {
-                claimant: protocol_fee_recipient,
-                claim_status_pubkey: protocol_claim_status_pubkey,
-                claim_status_bump: protocol_claim_status_bump,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: protocol_fee_amount,
-                proof: None,
-            }];
-
-            let (validator_claim_status_pubkey, validator_claim_status_bump) = Pubkey::find_program_address(
-                &[
-                    ClaimStatus::SEED,
-                    &stake_meta.validator_node_pubkey.to_bytes(),
-                    &tip_distribution_meta.tip_distribution_pubkey.to_bytes(),
-                ],
-                &JitoTipDistribution::id(),
-            );
-
-            tree_nodes.push(TreeNode {
-                claimant: stake_meta.validator_node_pubkey,
-                claim_status_pubkey: validator_claim_status_pubkey,
-                claim_status_bump: validator_claim_status_bump,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: validator_amount,
-                proof: None,
-            });
-
-            let total_delegated = stake_meta.total_delegated as u128;
-            tree_nodes.extend(
-                stake_meta.delegations
-                    .iter()
-                    .map(|delegation| {
-                        let amount_delegated = delegation.lamports_delegated as u128;
-                        let reward_amount = u128::div_ceil(
-                            amount_delegated * remaining_total_rewards as u128,
-                            total_delegated,
-                        );
-                        let reward_amount = u64::try_from(reward_amount)
-                            .map_err(|_| MerkleRootGeneratorError::CheckedMathError)?;
-
-                        println!(
-                            "Delegation: {}, Amount Delegated: {}, Reward Amount: {}",
-                            delegation.staker_pubkey, amount_delegated, reward_amount
-                        );
-
-                        let (claim_status_pubkey, claim_status_bump) = Pubkey::find_program_address(
-                            &[
-                                ClaimStatus::SEED,
-                                &delegation.staker_pubkey.to_bytes(),
-                                &tip_distribution_meta.tip_distribution_pubkey.to_bytes(),
-                            ],
-                            &JitoTipDistribution::id(),
-                        );
-                        Ok(TreeNode {
-                            claimant: delegation.staker_pubkey,
-                            claim_status_pubkey,
-                            claim_status_bump,
-                            staker_pubkey: delegation.staker_pubkey,
-                            withdrawer_pubkey: delegation.withdrawer_pubkey,
-                            amount: reward_amount,
-                            proof: None,
-                        })
-                    })
-                    .collect::<Result<Vec<TreeNode>, MerkleRootGeneratorError>>()?
-            );
-
-            Ok(Some(tree_nodes))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn hash(&self) -> Hash {
-        let mut hasher = Hasher::default();
-        hasher.hash(self.claimant.as_ref());
-        hasher.hash(self.amount.to_le_bytes().as_ref());
-        hasher.result()
     }
 }
 
@@ -821,322 +610,4 @@ pub fn read_json_from_file<T>(path: &PathBuf) -> serde_json::Result<T> where T: 
     let file = File::open(path).unwrap();
     let reader = BufReader::new(file);
     serde_json::from_reader(reader)
-}
-
-#[cfg(test)]
-mod tests {
-    use { super::*, jito_tip_distribution::merkle_proof };
-
-    #[test]
-    fn test_merkle_tree_verify() {
-        // Create the merkle tree and proofs
-        let tda = Pubkey::new_unique();
-        let (acct_0, acct_1) = (Pubkey::new_unique(), Pubkey::new_unique());
-        let claim_statuses = &[
-            (acct_0, tda),
-            (acct_1, tda),
-        ]
-            .iter()
-            .map(|(claimant, tda)| {
-                Pubkey::find_program_address(
-                    &[ClaimStatus::SEED, &claimant.to_bytes(), &tda.to_bytes()],
-                    &JitoTipDistribution::id()
-                )
-            })
-            .collect::<Vec<(Pubkey, u8)>>();
-        let tree_nodes = vec![
-            TreeNode {
-                claimant: acct_0,
-                claim_status_pubkey: claim_statuses[0].0,
-                claim_status_bump: claim_statuses[0].1,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: 151_507,
-                proof: None,
-            },
-            TreeNode {
-                claimant: acct_1,
-                claim_status_pubkey: claim_statuses[1].0,
-                claim_status_bump: claim_statuses[1].1,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: 176_624,
-                proof: None,
-            }
-        ];
-
-        // First the nodes are hashed and merkle tree constructed
-        let hashed_nodes: Vec<[u8; 32]> = tree_nodes
-            .iter()
-            .map(|n| n.hash().to_bytes())
-            .collect();
-        let mk = MerkleTree::new(&hashed_nodes[..], true);
-        let root = mk.get_root().expect("to have valid root").to_bytes();
-
-        // verify first node
-        let node = solana_program::hash::hashv(&[&[0u8], &hashed_nodes[0]]);
-        let proof = get_proof(&mk, 0);
-        assert!(merkle_proof::verify(proof, root, node.to_bytes()));
-
-        // verify second node
-        let node = solana_program::hash::hashv(&[&[0u8], &hashed_nodes[1]]);
-        let proof = get_proof(&mk, 1);
-        assert!(merkle_proof::verify(proof, root, node.to_bytes()));
-    }
-
-    #[tokio::test]
-    async fn test_new_from_stake_meta_collection_happy_path() {
-        let merkle_root_upload_authority = Pubkey::new_unique();
-        let (tda_0, tda_1) = (Pubkey::new_unique(), Pubkey::new_unique());
-        let stake_account_0 = Pubkey::new_unique();
-        let stake_account_1 = Pubkey::new_unique();
-        let stake_account_2 = Pubkey::new_unique();
-        let stake_account_3 = Pubkey::new_unique();
-        let staker_account_0 = Pubkey::new_unique();
-        let staker_account_1 = Pubkey::new_unique();
-        let staker_account_2 = Pubkey::new_unique();
-        let staker_account_3 = Pubkey::new_unique();
-        let validator_vote_account_0 = Pubkey::new_unique();
-        let validator_vote_account_1 = Pubkey::new_unique();
-        let validator_id_0 = Pubkey::new_unique();
-        let validator_id_1 = Pubkey::new_unique();
-
-        let stake_meta_collection = StakeMetaCollection {
-            stake_metas: vec![
-                StakeMeta {
-                    validator_vote_account: validator_vote_account_0,
-                    validator_node_pubkey: validator_id_0,
-                    maybe_tip_distribution_meta: Some(TipDistributionMeta {
-                        merkle_root_upload_authority,
-                        tip_distribution_pubkey: tda_0,
-                        total_tips: 1_900_122_111_000,
-                        validator_fee_bps: 100,
-                    }),
-                    delegations: vec![
-                        Delegation {
-                            stake_account_pubkey: stake_account_0,
-                            staker_pubkey: staker_account_0,
-                            withdrawer_pubkey: staker_account_0,
-                            lamports_delegated: 123_999_123_555,
-                        },
-                        Delegation {
-                            stake_account_pubkey: stake_account_1,
-                            staker_pubkey: staker_account_1,
-                            withdrawer_pubkey: staker_account_1,
-                            lamports_delegated: 144_555_444_556,
-                        }
-                    ],
-                    total_delegated: 1_555_123_000_333_454_000,
-                    commission: 100,
-                },
-                StakeMeta {
-                    validator_vote_account: validator_vote_account_1,
-                    validator_node_pubkey: validator_id_1,
-                    maybe_tip_distribution_meta: Some(TipDistributionMeta {
-                        merkle_root_upload_authority,
-                        tip_distribution_pubkey: tda_1,
-                        total_tips: 1_900_122_111_333,
-                        validator_fee_bps: 200,
-                    }),
-                    delegations: vec![
-                        Delegation {
-                            stake_account_pubkey: stake_account_2,
-                            staker_pubkey: staker_account_2,
-                            withdrawer_pubkey: staker_account_2,
-                            lamports_delegated: 224_555_444,
-                        },
-                        Delegation {
-                            stake_account_pubkey: stake_account_3,
-                            staker_pubkey: staker_account_3,
-                            withdrawer_pubkey: staker_account_3,
-                            lamports_delegated: 700_888_944_555,
-                        }
-                    ],
-                    total_delegated: 2_565_318_909_444_123,
-                    commission: 10,
-                }
-            ],
-            tip_distribution_program_id: Pubkey::new_unique(),
-            bank_hash: Hash::new_unique().to_string(),
-            epoch: 100,
-            slot: 2_000_000,
-        };
-
-        let merkle_tree_collection = GeneratedMerkleTreeCollection::new_from_stake_meta_collection(
-            stake_meta_collection.clone(),
-            300
-        ).await.unwrap();
-
-        assert_eq!(stake_meta_collection.epoch, merkle_tree_collection.epoch);
-        assert_eq!(stake_meta_collection.bank_hash, merkle_tree_collection.bank_hash);
-        assert_eq!(stake_meta_collection.slot, merkle_tree_collection.slot);
-        assert_eq!(
-            stake_meta_collection.stake_metas.len(),
-            merkle_tree_collection.generated_merkle_trees.len()
-        );
-
-        let protocol_fee_recipient = Pubkey::find_program_address(
-            &[b"protocol_fee", &(0u64).to_le_bytes()],
-            &stake_meta_collection.tip_distribution_program_id
-        ).0;
-
-        let claim_statuses = &[
-            (protocol_fee_recipient, tda_0),
-            (validator_vote_account_0, tda_0),
-            (stake_account_0, tda_0),
-            (stake_account_1, tda_0),
-            (protocol_fee_recipient, tda_1),
-            (validator_vote_account_1, tda_1),
-            (stake_account_2, tda_1),
-            (stake_account_3, tda_1),
-        ]
-            .iter()
-            .map(|(claimant, tda)| {
-                Pubkey::find_program_address(
-                    &[ClaimStatus::SEED, &claimant.to_bytes(), &tda.to_bytes()],
-                    &JitoTipDistribution::id()
-                )
-            })
-            .collect::<Vec<(Pubkey, u8)>>();
-
-        let tree_nodes = vec![
-            TreeNode {
-                claimant: protocol_fee_recipient,
-                claim_status_pubkey: claim_statuses[0].0,
-                claim_status_bump: claim_statuses[0].1,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: 57_003_663_330, // 3% of 1_900_122_111_000
-                proof: None,
-            },
-            TreeNode {
-                claimant: validator_id_0, // Changed from validator_vote_account_0 to validator_id_0
-                claim_status_pubkey: claim_statuses[1].0,
-                claim_status_bump: claim_statuses[1].1,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: 19_001_221_110,
-                proof: None,
-            },
-            TreeNode {
-                claimant: staker_account_0, // Use staker_account instead of stake_account
-                claim_status_pubkey: claim_statuses[2].0,
-                claim_status_bump: claim_statuses[2].1,
-                staker_pubkey: staker_account_0,
-                withdrawer_pubkey: staker_account_0,
-                amount: 145_447, // Update to match actual amount
-                proof: None,
-            },
-            TreeNode {
-                claimant: staker_account_1, // Use staker_account instead of stake_account
-                claim_status_pubkey: claim_statuses[3].0,
-                claim_status_bump: claim_statuses[3].1,
-                staker_pubkey: staker_account_1,
-                withdrawer_pubkey: staker_account_1,
-                amount: 169_559, // Update to match actual amount
-                proof: None,
-            }
-        ];
-
-        let hashed_nodes: Vec<[u8; 32]> = tree_nodes
-            .iter()
-            .map(|n| n.hash().to_bytes())
-            .collect();
-        let merkle_tree = MerkleTree::new(&hashed_nodes[..], true);
-        let gmt_0 = GeneratedMerkleTree {
-            tip_distribution_account: tda_0,
-            merkle_root_upload_authority,
-            merkle_root: *merkle_tree.get_root().unwrap(),
-            tree_nodes,
-            max_total_claim: stake_meta_collection.stake_metas[0]
-                .clone()
-                .maybe_tip_distribution_meta.unwrap().total_tips,
-            max_num_nodes: 4,
-        };
-
-        let tree_nodes = vec![
-            TreeNode {
-                claimant: protocol_fee_recipient,
-                claim_status_pubkey: claim_statuses[4].0,
-                claim_status_bump: claim_statuses[4].1,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: 57_003_663_339,
-                proof: None,
-            },
-            TreeNode {
-                claimant: validator_id_1,
-                claim_status_pubkey: claim_statuses[5].0,
-                claim_status_bump: claim_statuses[5].1,
-                staker_pubkey: Pubkey::default(),
-                withdrawer_pubkey: Pubkey::default(),
-                amount: 38_002_442_226,
-                proof: None,
-            },
-            TreeNode {
-                claimant: staker_account_2,
-                claim_status_pubkey: claim_statuses[6].0,
-                claim_status_bump: claim_statuses[6].1,
-                staker_pubkey: staker_account_2,
-                withdrawer_pubkey: staker_account_2,
-                amount: 158_011, // Updated from 163_000
-                proof: None,
-            },
-            TreeNode {
-                claimant: staker_account_3,
-                claim_status_pubkey: claim_statuses[7].0,
-                claim_status_bump: claim_statuses[7].1,
-                staker_pubkey: staker_account_3,
-                withdrawer_pubkey: staker_account_3,
-                amount: 493_188_526, // Updated from 508_762_900
-                proof: None,
-            }
-        ];
-        let hashed_nodes: Vec<[u8; 32]> = tree_nodes
-            .iter()
-            .map(|n| n.hash().to_bytes())
-            .collect();
-        let merkle_tree = MerkleTree::new(&hashed_nodes[..], true);
-        let gmt_1 = GeneratedMerkleTree {
-            tip_distribution_account: tda_1,
-            merkle_root_upload_authority,
-            merkle_root: *merkle_tree.get_root().unwrap(),
-            tree_nodes,
-            max_total_claim: stake_meta_collection.stake_metas[1]
-                .clone()
-                .maybe_tip_distribution_meta.unwrap().total_tips,
-            max_num_nodes: 4,
-        };
-
-        let expected_generated_merkle_trees = vec![gmt_0, gmt_1];
-        let actual_generated_merkle_trees = merkle_tree_collection.generated_merkle_trees;
-        expected_generated_merkle_trees.iter().for_each(|expected_gmt| {
-            let actual_gmt = actual_generated_merkle_trees
-                .iter()
-                .find(|gmt| gmt.tip_distribution_account == expected_gmt.tip_distribution_account)
-                .unwrap();
-
-            println!("Expected nodes:");
-            for node in &expected_gmt.tree_nodes {
-                println!("Claimant: {}, Amount: {}", node.claimant, node.amount);
-            }
-            println!("\nActual nodes:");
-            for node in &actual_gmt.tree_nodes {
-                println!("Claimant: {}, Amount: {}", node.claimant, node.amount);
-            }
-
-            assert_eq!(expected_gmt.max_num_nodes, actual_gmt.max_num_nodes);
-            assert_eq!(expected_gmt.max_total_claim, actual_gmt.max_total_claim);
-            assert_eq!(expected_gmt.tip_distribution_account, actual_gmt.tip_distribution_account);
-            assert_eq!(expected_gmt.tree_nodes.len(), actual_gmt.tree_nodes.len());
-            expected_gmt.tree_nodes.iter().for_each(|expected_tree_node| {
-                let actual_tree_node = actual_gmt.tree_nodes
-                    .iter()
-                    .find(|tree_node| tree_node.claimant == expected_tree_node.claimant)
-                    .unwrap();
-                assert_eq!(expected_tree_node.amount, actual_tree_node.amount);
-            });
-            assert_eq!(expected_gmt.merkle_root, actual_gmt.merkle_root);
-        });
-    }
 }
